+++ directory_components/forms/fields.py
@@ -5,19 +5,55 @@ from directory_components.forms import widgets
 
 
 __all__ = [
+    'DirectoryComponentsFieldMixin',
+    'BindNestedFormMixin',
+    'DirectoryComponentsBoundField',
+    'field_factory',
+    'PaddedCharField',
+    'RadioNested',
     'BooleanField',
     'CharField',
     'ChoiceField',
     'DateField',
-    'DirectoryComponentsBoundField',
-    'DirectoryComponentsFieldMixin',
+    'DateTimeField',
+    'DecimalField',
+    'DurationField',
     'EmailField',
-    'field_factory',
+    'FileField',
+    'FilePathField',
+    'FloatField',
+    'GenericIPAddressField',
+    'ImageField',
     'IntegerField',
     'MultipleChoiceField',
-    'PaddedCharField',
+    'RegexField',
+    'SlugField',
+    'TimeField',
+    'TypedChoiceField',
+    'TypedMultipleChoiceField',
     'URLField',
+    'UUIDField',
 ]
+
+
+class BindNestedFormMixin:
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        for name, field in self.fields.items():
+            if isinstance(field, RadioNested):
+                nested_form = field.nested_form_class(*args, **kwargs)
+                # require the nested fields to be provided if the parent field is checked
+                if field.coerce(self[name].data) != field.nested_form_choice:
+                    for item in nested_form.fields.values():
+                        item.required = False
+                field.bind_nested_form(nested_form)
+
+    def clean(self):
+        super().clean()
+        for field_name in list(self.cleaned_data.keys()):
+            field = self.fields[field_name]
+            if isinstance(field, RadioNested) and field.nested_form.is_valid():
+                self.cleaned_data.update(field.nested_form.cleaned_data)
 
 
 class DirectoryComponentsBoundField(BoundField):
@@ -38,13 +74,16 @@ class DirectoryComponentsBoundField(BoundField):
 class DirectoryComponentsFieldMixin:
 
     def __init__(self, container_css_classes='form-group', *args, **kwargs):
-        self.container_css_classes = container_css_classes
         super().__init__(*args, **kwargs)
         if not hasattr(self.widget, 'css_class_name'):
-            self.widget.attrs['class'] = (
-                self.widget.attrs.get('class', '') + ' form-control'
-            )
+            self.widget.attrs['class'] = self.widget.attrs.get('class', '') + ' form-control'
         self.label_suffix = ''
+        self._container_css_classes = container_css_classes
+
+    @property
+    def container_css_classes(self):
+        widget_class = getattr(self.widget, 'container_css_classes', '')
+        return f'{self._container_css_classes} {widget_class}'
 
     def get_bound_field(self, form, field_name):
         return DirectoryComponentsBoundField(form, self, field_name)
@@ -56,23 +95,37 @@ def field_factory(base_class):
 
 
 CharField = field_factory(forms.CharField)
-EmailField = field_factory(forms.EmailField)
-URLField = field_factory(forms.URLField)
 ChoiceField = field_factory(forms.ChoiceField)
 DateField = field_factory(forms.DateField)
+DateTimeField = field_factory(forms.DateTimeField)
+DecimalField = field_factory(forms.DecimalField)
+DurationField = field_factory(forms.DurationField)
+EmailField = field_factory(forms.EmailField)
+FileField = field_factory(forms.FileField)
+FilePathField = field_factory(forms.FilePathField)
+FloatField = field_factory(forms.FloatField)
+GenericIPAddressField = field_factory(forms.GenericIPAddressField)
+ImageField = field_factory(forms.ImageField)
 IntegerField = field_factory(forms.IntegerField)
 MultipleChoiceField = field_factory(forms.MultipleChoiceField)
+RegexField = field_factory(forms.RegexField)
+SlugField = field_factory(forms.SlugField)
+TimeField = field_factory(forms.TimeField)
+TypedChoiceField = field_factory(forms.TypedChoiceField)
+TypedMultipleChoiceField = field_factory(forms.TypedMultipleChoiceField)
+URLField = field_factory(forms.URLField)
+UUIDField = field_factory(forms.UUIDField)
 
 
 class BooleanField(DirectoryComponentsFieldMixin, forms.BooleanField):
-    def __init__(self, label='', help_text='', widget=None, *args, **kwargs):
-        if widget is None:
-            widget = widgets.CheckboxWithInlineLabel()
-        if isinstance(widget, widgets.CheckboxWithInlineLabel):
-            widget.label = label
-            widget.help_text = help_text
-            label = ''
-        super().__init__(label=label, widget=widget, *args, **kwargs)
+    widget = widgets.CheckboxWithInlineLabel
+
+    def __init__(self, label='', help_text='', *args, **kwargs):
+        super().__init__(label=label, *args, **kwargs)
+        if isinstance(self.widget, widgets.CheckboxWithInlineLabel):
+            self.widget.label = label
+            self.widget.help_text = help_text
+            self.label = ''
 
 
 class PaddedCharField(CharField):
@@ -85,3 +138,28 @@ class PaddedCharField(CharField):
         if value not in self.empty_values:
             return value.rjust(self.max_length, self.fillchar)
         return value
+
+
+class RadioNested(TypedChoiceField):
+    MESSAGE_FORM_MIXIN = 'This field requires the form to use BindNestedFormMixin'
+    widget = widgets.RadioNestedWidget
+
+    def __init__(self, nested_form_class=None, nested_form_choice=True, *args, **kwargs):
+        self.nested_form_class = nested_form_class
+        self.nested_form_choice = nested_form_choice
+        super().__init__(*args, **kwargs)
+        self.widget.nested_form_choice = nested_form_choice
+
+    def bind_nested_form(self, form):
+        self.nested_form = form
+        self.widget.bind_nested_form(form)
+
+    def validate(self, value):
+        super().validate(value)
+        if value and not self.nested_form.is_valid():
+            # trigger the form to mark the field as invalid. the nested form will then render the real errors
+            raise forms.ValidationError(message='')
+
+    def get_bound_field(self, form, field_name):
+        assert isinstance(form, BindNestedFormMixin), self.MESSAGE_FORM_MIXIN
+        return super().get_bound_field(form, field_name)