+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 35.4.0
+[Full Changelog](https://github.com/uktrade/directory-components/pull/309)
+### Implemented enhancements
+- Added RadioNestedField
+- Added TextInputWithSubmitButton widget
+- Added SelectMultipleAutocomplete widget
+
 ## 35.3.1
 [Full Changelog](https://github.com/uktrade/directory-components/pull/311)
 ### Fixed bugs
@@ -9,6 +16,7 @@
 [Full Changelog](https://github.com/uktrade/directory-components/pull/308)
 ### Implemented enhancements
 - remove arrow from hero cta button
+
 ## 35.2.1
 [Full Changelog](https://github.com/uktrade/directory-components/pull/307)
 ### Fixed bugs